# (#3921) - small optimizations to merge.js