--- a/lib/merge.js
+++ b/lib/merge.js
@@ -1,6 +1,4 @@
 'use strict';
-var extend = require('pouchdb-extend');
-
 
 // for a better overview of what this is doing, read:
 // https://github.com/apache/couchdb/blob/master/src/couchdb/couch_key_tree.erl
@@ -27,6 +25,20 @@ function binarySearch(arr, item, comparator) {
     }
   }
   return low;
+}
+
+function sortByPos(a, b) {
+  return a.pos - b.pos;
+}
+
+function sortByDeletedThenPosThenId(a, b) {
+  if (a.deleted !== b.deleted) {
+    return a.deleted > b.deleted ? 1 : -1;
+  }
+  if (a.pos !== b.pos) {
+    return b.pos - a.pos;
+  }
+  return a.id < b.id ? 1 : -1;
 }
 
 // assuming the arr is sorted, insert the item in the proper place
@@ -105,7 +117,8 @@ function doMerge(tree, path, dontExpand) {
     return {tree: [path], conflicts: 'new_leaf'};
   }
 
-  tree.forEach(function (branch) {
+  for (var i = 0, len = tree.length; i < len; i++) {
+    var branch = tree[i];
     if (branch.pos === path.pos && branch.ids[0] === path.ids[0]) {
       // Paths start at the same position and have the same root, so they need
       // merged
@@ -135,11 +148,15 @@ function doMerge(tree, path, dontExpand) {
           }
           continue;
         }
-        /*jshint loopfunc:true */
-        item.ids[2].forEach(function (el, idx) {
-          trees.push(
-            {ids: el, diff: item.diff - 1, parent: item.ids, parentIdx: idx});
-        });
+        var elements = item.ids[2];
+        for (var j = 0, elementsLen = elements.length; j < elementsLen; j++) {
+          trees.push({
+            ids: elements[j],
+            diff: item.diff - 1,
+            parent: item.ids,
+            parentIdx: j
+          });
+        }
       }
 
       var el = candidateParents[0];
@@ -156,16 +173,14 @@ function doMerge(tree, path, dontExpand) {
     } else {
       restree.push(branch);
     }
-  });
+  }
 
   // We didnt find
   if (!merged) {
     restree.push(path);
   }
 
-  restree.sort(function (a, b) {
-    return a.pos - b.pos;
-  });
+  restree.sort(sortByPos);
 
   return {
     tree: restree,
@@ -194,9 +209,6 @@ function stem(tree, depth) {
 var PouchMerge = {};
 
 PouchMerge.merge = function (tree, path, depth) {
-  // Ugh, nicer way to not modify arguments in place?
-  tree = extend(true, [], tree);
-  path = extend(true, {}, path);
   var newTree = doMerge(tree, path);
   return {
     tree: stem(newTree.tree, depth),
@@ -217,15 +229,7 @@ PouchMerge.winningRev = function (metadata) {
       leafs.push({pos: pos, id: id, deleted: !!opts.deleted});
     }
   });
-  leafs.sort(function (a, b) {
-    if (a.deleted !== b.deleted) {
-      return a.deleted > b.deleted ? 1 : -1;
-    }
-    if (a.pos !== b.pos) {
-      return b.pos - a.pos;
-    }
-    return a.id < b.id ? 1 : -1;
-  });
+  leafs.sort(sortByDeletedThenPosThenId);
 
   return leafs[0].pos + '-' + leafs[0].id;
 };
@@ -257,10 +261,10 @@ PouchMerge.collectLeaves = function (revs) {
       leaves.push({rev: pos + "-" + id, pos: pos, opts: opts});
     }
   });
-  leaves.sort(function (a, b) {
-    return b.pos - a.pos;
-  });
-  leaves.forEach(function (leaf) { delete leaf.pos; });
+  leaves.sort(sortByPos).reverse();
+  for (var i = 0, len = leaves.length; i < len; i++) {
+    delete leaves[i].pos;
+  }
   return leaves;
 };
 
@@ -271,11 +275,12 @@ PouchMerge.collectConflicts = function (metadata) {
   var win = PouchMerge.winningRev(metadata);
   var leaves = PouchMerge.collectLeaves(metadata.rev_tree);
   var conflicts = [];
-  leaves.forEach(function (leaf) {
+  for (var i = 0, len = leaves.length; i < len; i++) {
+    var leaf = leaves[i];
     if (leaf.rev !== win && !leaf.opts.deleted) {
       conflicts.push(leaf.rev);
     }
-  });
+  }
   return conflicts;
 };
 
@@ -286,11 +291,11 @@ PouchMerge.rootToLeaf = function (tree) {
     history.push({id: id, opts: opts});
     if (isLeaf) {
       var rootPos = pos + 1 - history.length;
-      paths.unshift({pos: rootPos, ids: history});
+      paths.push({pos: rootPos, ids: history});
     }
     return history;
   });
-  return paths;
+  return paths.reverse();
 };