According to performance profiling I've been doing, the big boost here should come from not calling `extend()`, which it turned out was unnecessary anyway. I also went through and removed as many functions-within-functions as I could, and migrated everything from `forEach()` to regular `for` loops.

When this is green, I'll provide some perf numbers to justify it. Looking at the Chrome profiler, though, this should help a lot.