@nolanlawson 2015-08-09
Okay, so AFAICT, our `merge()` function now actually mutates the object it is given as input. But it doesn't seem to matter, because whenever we do a merge, it's because we're going to update the single `metadata` object in the database anyway. So mutating the input tree is actually just fine, and in fact much better for performance than calling `extend()` all over the place.

However, this would normally break the `test.merge.js` unit test, so I modified it to redefine the constants every time. This is kinda risky, but since our integration tests are passing, I feel confident doing this.