@nolanlawson 2015-08-09
Performance boost for the `bulk-inserts` test (10 iterations in browser, 100 in Node for better granularity)

* iojs 2.5 (leveldb): 36855ms -> 36822ms
* Chrome 44 (idb): 15179ms -> 14471ms
* Chrome 11 (websql): 32476ms -> 30788ms

I admit, I wish the boost were bigger in Node. :( It's an improvement, but not a huge one.