@daleharvey 2015-08-10
Following up on the coverage failure in https://github.com/pouchdb/pouchdb/issues/4142

Do you still think its worth it? modifying args in place is the default behaviour of javascript, so its arguably not confusing, I do think its terrible confusing behaviour and it turns this from a decoupled functional api to one that could be potentially confusing, however we are the only consumers. 

Having mixed feelings really