@nolanlawson 2015-08-11
I have mixed feelings as well. I think I will split this into the micro-optimizations (which are fairly uncontroversial) and the big optimization (removing `extend`). Then I want some better measurements about how much exactly this is improving.

When I profile I consistently see `extendInner` very high on the list. It might not figure highly into the raw speed numbers of our perf test, but in terms of what is taking up the most *JavaScript* time it is quite high, which means it's the kind of thing that could be blocking the DOM. So yeah, I think it's imprortant, but we should be cautious about adding such stuff.