@nolanlawson 2015-08-12
Closing in favor of two new PRs